Aggregates minimax algorithm with computerplayer and removes minimax object.
