+++ src/test/java/Core/Players/PlayerComputerTest.java
@@ -2,18 +2,17 @@ package Core.Players;
 
 import Core.Grid;
 import Core.Mark;
-import Core.Minimax;
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
 
-public class MinimaxTest {
+public class PlayerComputerTest {
 
 
     @Test
     public void minimaxTest1() {
         Grid grid = new Grid();
-        Minimax minimax = new Minimax(grid, Mark.playerTwoMark, Mark.playerOneMark);
+        Player computer = new PlayerComputer("Computer", Mark.playerTwoMark, Mark.playerOneMark);
 
 //         O X O
 //         O O X
@@ -27,13 +26,13 @@ public class MinimaxTest {
         grid.markSquare(5, Mark.playerOneMark);
         grid.markSquare(6, Mark.playerOneMark);
 
-        assertEquals(8, minimax.findBestMove());
+        assertEquals(8, computer.getInput(grid));
     }
 
     @Test
     public void minimaxTest2() {
         Grid grid = new Grid();
-        Minimax minimax = new Minimax(grid, Mark.playerTwoMark, Mark.playerOneMark);
+        Player computer = new PlayerComputer("Computer", Mark.playerTwoMark, Mark.playerOneMark);
 
 //         O X O
 //         _ _ X
@@ -47,13 +46,13 @@ public class MinimaxTest {
         grid.markSquare(7, Mark.playerTwoMark);
         grid.markSquare(8, Mark.playerTwoMark);
 
-        assertEquals(4, minimax.findBestMove());
+        assertEquals(4, computer.getInput(grid));
     }
 
     @Test
     public void minimaxTest3() {
         Grid grid = new Grid();
-        Minimax minimax = new Minimax(grid, Mark.playerTwoMark, Mark.playerOneMark);
+        Player computer = new PlayerComputer("Computer", Mark.playerTwoMark, Mark.playerOneMark);
 
 //         O _ X
 //         X _ _
@@ -66,13 +65,13 @@ public class MinimaxTest {
         grid.markSquare(7, Mark.playerTwoMark);
         grid.markSquare(8, Mark.playerTwoMark);
 
-        assertEquals(4, minimax.findBestMove());
+        assertEquals(4, computer.getInput(grid));
     }
 
     @Test
     public void minimaxTest4() {
         Grid grid = new Grid();
-        Minimax minimax = new Minimax(grid, Mark.playerTwoMark, Mark.playerOneMark);
+        Player computer = new PlayerComputer("Computer", Mark.playerTwoMark, Mark.playerOneMark);
 
 //         X X _
 //         _ X _
@@ -84,13 +83,13 @@ public class MinimaxTest {
         grid.markSquare(6, Mark.playerTwoMark);
         grid.markSquare(8, Mark.playerTwoMark);
 
-        assertEquals(7, minimax.findBestMove());
+        assertEquals(7, computer.getInput(grid));
     }
 
     @Test
     public void minimaxTest5() {
         Grid grid = new Grid();
-        Minimax minimax = new Minimax(grid, Mark.playerTwoMark, Mark.playerOneMark);
+        Player computer = new PlayerComputer("Computer", Mark.playerTwoMark, Mark.playerOneMark);
 
 //         X _ O
 //         O _ _
@@ -103,13 +102,13 @@ public class MinimaxTest {
         grid.markSquare(7, Mark.playerOneMark);
         grid.markSquare(8, Mark.playerOneMark);
 
-        assertEquals(4, minimax.findBestMove());
+        assertEquals(4, computer.getInput(grid));
     }
 
     @Test
     public void minimaxTest6WithDepth() {
         Grid grid = new Grid();
-        Minimax minimax = new Minimax(grid, Mark.playerTwoMark, Mark.playerOneMark);
+        Player computer = new PlayerComputer("Computer", Mark.playerTwoMark, Mark.playerOneMark);
 
 //         _ X _
 //         _ _ X
@@ -121,13 +120,13 @@ public class MinimaxTest {
         grid.markSquare(7, Mark.playerTwoMark);
         grid.markSquare(8, Mark.playerOneMark);
 
-        assertEquals(2, minimax.findBestMove());
+        assertEquals(2, computer.getInput(grid));
     }
 
     @Test
     public void minimaxTest7WithDepthWithSwitchedMarks() {
         Grid grid = new Grid();
-        Minimax minimax = new Minimax(grid, Mark.playerOneMark, Mark.playerTwoMark);
+        Player computer = new PlayerComputer("Computer", Mark.playerTwoMark, Mark.playerOneMark);
 
 //         O O X
 //         _ X _
@@ -138,6 +137,6 @@ public class MinimaxTest {
         grid.markSquare(2, Mark.playerTwoMark);
         grid.markSquare(4, Mark.playerTwoMark);
 
-        assertEquals(6, minimax.findBestMove());
+        assertEquals(6, computer.getInput(grid));
     }
 }
